feat: 상세 조회 완료
